requirements.txt
streamlit==1.25.0
Pillow==9.4.0
pytesseract==0.3.10
opencv-python==4.8.0.76
numpy==1.25.0
setuptools>=65.5.0
